Rename folder without tx shorthand
