+++ packages/backend/src/Application.ts
@@ -20,7 +20,7 @@ import { Clock } from './core/Clock'
 import { EventUpdater } from './core/events/EventUpdater'
 import { PriceUpdater } from './core/PriceUpdater'
 import { ReportUpdater } from './core/reports/ReportUpdater'
-import { RpcTransactionUpdater } from './core/tx-count/RpcTransactionUpdater'
+import { RpcTransactionUpdater } from './core/transaction-count/RpcTransactionUpdater'
 import { CoingeckoQueryService } from './peripherals/coingecko/CoingeckoQueryService'
 import { AggregateReportRepository } from './peripherals/database/AggregateReportRepository'
 import { BalanceRepository } from './peripherals/database/BalanceRepository'
@@ -62,7 +62,10 @@ export class Application {
       logger,
     )
     const eventRepository = new EventRepository(database, logger)
-    const txCountRepository = new TransactionCountRepository(database, logger)
+    const transactionCountRepository = new TransactionCountRepository(
+      database,
+      logger,
+    )
 
     const http = new HttpClient()
 
@@ -146,7 +149,7 @@ export class Application {
     // TODO buildUpdaterForAll
     const optimismRpcTransactionUpdater = new RpcTransactionUpdater(
       optimismClient,
-      txCountRepository,
+      transactionCountRepository,
       clock,
       logger,
       ProjectId('optimism'),
